Move helper function to a separate file.
